requirements.txt
pytorch-lightning==2.4.0
torchvision==0.19.0
wandb==0.17.6
omegaconf==2.3.0
pandas==2.2.2
submitit==1.5.1
numpy==2.0.1
timm==1.0.8
scikit-learn==1.5.1

# data download
selenium==4.23.1
python-dotenv==1.0.1
webdriver_manager==4.0.2